backend/charter/test_full.py
#!/usr/bin/env python3
"""
Full integration test for Charter agent after Lambda deployment
This will be used to test the deployed Lambda function
"""

import os
import json
import boto3
from dotenv import load_dotenv

load_dotenv(override=True)

def test_charter_lambda():
    """Test the deployed Charter Lambda function"""
    print("Charter Lambda integration test - to be implemented after deployment")
    print("This will test the actual deployed Lambda function via boto3 invoke")
    pass

if __name__ == "__main__":
    test_charter_lambda()